created a solution file with two separate csproj ( one for demo with the main and another for the library itself )
