+++ ConvexHull.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25123.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConvexHull.Demo", "ConvexHull.Demo\ConvexHull.Demo.csproj", "{A58BDAC7-5DC1-46C3-8196-15EB6F4A7CF9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConvexHull", "ConvexHull\ConvexHull.csproj", "{DC956E5C-5FAB-481E-BA9F-3BEB4E7E6B20}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A58BDAC7-5DC1-46C3-8196-15EB6F4A7CF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A58BDAC7-5DC1-46C3-8196-15EB6F4A7CF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A58BDAC7-5DC1-46C3-8196-15EB6F4A7CF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A58BDAC7-5DC1-46C3-8196-15EB6F4A7CF9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DC956E5C-5FAB-481E-BA9F-3BEB4E7E6B20}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DC956E5C-5FAB-481E-BA9F-3BEB4E7E6B20}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DC956E5C-5FAB-481E-BA9F-3BEB4E7E6B20}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DC956E5C-5FAB-481E-BA9F-3BEB4E7E6B20}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
